requirements.txt
flask==3.1.0
flask-login==0.6.3
requests==2.32.3
urllib3==2.3.0
beautifulsoup4==4.13.3
lxml==5.3.0
werkzeug==3.1.3
reportlab==4.3.1
schedule==1.2.2
python-dotenv==1.0.0
